dashboard/api/requirements.txt
# Dashboard API Dependencies

fastapi>=0.110.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
httpx>=0.27.0
pydantic>=2.6.0

# Google Cloud
google-cloud-firestore>=2.16.0
google-cloud-storage>=2.14.0
google-cloud-secret-manager>=2.18.0

# Encryption
cryptography>=42.0.0

# Session/Auth
python-jose[cryptography]>=3.3.0
